Fix 10x energy/runtime over-report; absorb idle-socket reconnect blip
Two fixes for the AR12HSFS:

1. lifetime_energy / operating_time were 10x too high. AC_ADD2_USEDPOWER
   and AC_ADD2_USEDTIME are fixed-point in tenths (like the capability
   registers and the documented AC_ADD2_USEDWATT raw/10 convention), but
   were read raw. So 6404 -> 640.4 kWh and 10820 -> 1082.0 h. Confirmed
   against ground truth: on 2026-06-23 the unit ran 08:33-18:48 (~10.1 h
   measured) and the runtime register logged +100 -> 10.0 h at /10, with
   ~0.5 kW average draw. Raw would claim 100 operating-hours in one
   calendar day (impossible). Divide both by 10; PROTOCOL.md updated.

2. The AC silently drops its idle TLS socket between 30s polls, so the
   first read on the stale connection times out and the entity flapped to
   "unavailable" for a poll on every reconnect. The coordinator now
   reconnects and retries once before surfacing the failure.

+++ custom_components/samsung_2878/sensor.py
@@ -194,7 +194,7 @@ class UsedPowerSensor(CoordinatorEntity[Samsung2878Coordinator], SensorEntity):
         self._attr_device_info = device_info
 
     @property
-    def native_value(self) -> int | None:
+    def native_value(self) -> float | None:
         """Return lifetime energy usage in kWh."""
         return self.coordinator.data.used_power
 
@@ -220,7 +220,7 @@ class UsedTimeSensor(CoordinatorEntity[Samsung2878Coordinator], SensorEntity):
         self._attr_device_info = device_info
 
     @property
-    def native_value(self) -> int | None:
+    def native_value(self) -> float | None:
         """Return operating time in hours."""
         return self.coordinator.data.used_time
